+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2025-11-25
+
 ### Added
 - Add `stud help <command>` to display detailed help for a specific command, with support for command aliases [SCI-7]
 - Add `--info` (`-i`) option to `stud update` command to preview changelog without downloading [SCI-6]